pyproject.toml
[project]
name = "postgrest-jwt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
]
